Common entity moved to libs+fixed imports
