+++ apps/core/src/auth/strategies/jwt.strategy.ts
@@ -4,7 +4,7 @@ import {Injectable, UnauthorizedException} from '@nestjs/common';
 import {ConfigService} from '@nestjs/config';
 import {InjectRepository} from '@nestjs/typeorm';
 import {Repository} from 'typeorm';
-import {EnvironmentVariables} from '../../common/EnvironmentVariables';
+import {EnvironmentVariables} from '@ezyfs/common/types';
 import {PayloadInterface} from '../dto/payload.interface';
 import {User} from '../../user/entities/user.entity';
