+++ apps/core/src/crypto/entities/AsymKey.entity.ts
@@ -1,5 +1,5 @@
 import {ObjectType, Field, Int} from '@nestjs/graphql';
-import {TimestampEntites} from '../../generics/timestamp.entity';
+import {TimestampEntites} from '@ezyfs/common/entities/timestamp.entity';
 import {User} from '../../user/entities/user.entity';
 import {
   Column,
