+++ libs/common/src/entities/timestamp.entity.ts
@@ -1,0 +1,19 @@
+import {Field, ObjectType} from '@nestjs/graphql';
+import {CreateDateColumn, UpdateDateColumn, DeleteDateColumn} from 'typeorm';
+
+@ObjectType()
+export class TimestampEntites {
+  @Field()
+  @CreateDateColumn({
+    update: false,
+  })
+  createdAt: Date;
+
+  @Field()
+  @UpdateDateColumn()
+  updatedAt: Date;
+
+  @Field()
+  @DeleteDateColumn()
+  deletedAt: Date;
+}
