+++ libs/common/src/types/EnvironmentVariables.ts
